Fix XECS-655 - Sipxacd dumpes the core when the ACD server is Activated frequently.
Applied patch from Huijun Yang.  He writes:

Here is the root cause of the issue: When the ProvisingAgent is in the 
process of being destructed, requests come in with a pointer to the 
ProvisingAgent, and caused a invalid memory access, thus caused core dump. 

The fix is to close the window that to allow ProvisingAgent being accessed 
while it is being destructed.


Thanks Huijun!

--Woof!



git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11493 ab1d8caa-1f67-47f1-9e81-24633a41865c