Fix for XECS-715: User without Voicemail permission but with     Autoattendent permission, is able to receive the voicemail
This problem was introduced in r11356 / XECS-402.  The logic in 
validateMailbox was updated to pass the permission checks if
either VoiceMail or the AutoAttendant permission was available.
This change was needed to allow the end-user to record their
name.

validateMailbox now looks at the requested/required permissions
instead of either.  The retrieve/login CGIs now require either
VoiceMail or AutoAttendant while all other CGIs now require
Voicemail (fixing this issue).



git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11543 ab1d8caa-1f67-47f1-9e81-24633a41865c