XCF-2296 make sure that DB patches are compatible with Postgres 7.4
RHEL/4 is using Postgres 7.4 which does not support adding a table column and setting a default for a column in a single SQL statement.
Because of that when adding new column we need to write code that looks like this.

alter table t add column c boolean;
alter table t alter column c set default false;


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11997 ab1d8caa-1f67-47f1-9e81-24633a41865c