XECS-1078 Allow configuring AA transfer to any valid SIP URI
Changed the menu item from "Transfer to Extension" to a more generic "Transfer Out". It can be either an extension, a sip URI, or anything else reachable via the dial plans (including external numbers). For example, to transfer to an extension's mailbox, "8xxx" could be used, assuming the default dial plan vm prefix "8".

If a domain name is left off, sipXconfig will insert "this" domain name into the appropriate place in the VXML.

Values entered here bypass the CGI check for a valid extension. Invalid destinations will just cause the caller to be hung up on (due to blind transfer). It is hoped that the new authorization scheme is enough to prevent abuse of this feature.

Please note that extension values entered by DTMF digits on the AA do NOT bypass the CGI check. We do NOT want to allow AA callers to be able to dial anything reachable by the dial plans. Those numbers are still vetted for valid extensions and alias's before being allowed to be dialed.

Patch from woof! - thanks!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12000 ab1d8caa-1f67-47f1-9e81-24633a41865c