Replace standard jwebunit libraries with sipx build version.
This is hopefully a temporary hack. Because compiling dojo does not really work from htmlunit
we need to add several features and fixes that are not available in standard version.
- committing a form wihout specifying submit button - used by submit link
- setting hidden fields - used by submit link
- better checkbox support - jwebunit seems to have problems with supporting checkboxes which
   identifiers have ':' (colon) in them

I will make our patches public, but apart from checkbox changes there is little hope they would
be considered for inclusion in jwebunit proper.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12264 ab1d8caa-1f67-47f1-9e81-24633a41865c