XCF-2614 replacing replication.cgi with XML/RPC calls
New interfaces FileApi and ImdbApi defined to match XML/RPC interface of sipXsupervisor.
DataSetGenerator objects do not generate XML any more. Instead they build a list of maps; each map
represents a single item (row) in IMDB table.  XML generation is preserved for now for logging and
debugging purposes (see deprecated generateXml method)

Most tests are rewritten not to use XML comparison as a way of testing dataset generation.

File replication requires that sipXconfig provides a full path of the replicated file. For now we
are using {sipx.etc} property, but a better solution is needed since not all the files are in
/etc/sipxpbx directory.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12965 ab1d8caa-1f67-47f1-9e81-24633a41865c