XCF-2679 limit forwarding for request initiated by sipXconfig
sipXconfig send SIP requests to implement click-to-call and check-sync notifications
Those requests should not be routed to voicemail or participate if call forwarding: they are really
directed to specific phones and not to users, but we do not have a good way of sending requests to
phones.
Added sipx specific URL parameters: 'sipx-noroute' and 'sipx-userforward'

Patch from our very own Damian Krzeminski - thanks!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13583 ab1d8caa-1f67-47f1-9e81-24633a41865c