XCF-2809 Removed sipXconfig dependency on SIPXCHANGE_REALM and SIPXCHANGE_DOMAIN_NAME from config.defs
Modified sipXconfig to read realm and doman and domain info from config.defs file when system is first started. This file is created by sipx-setup.  After this initial read of the file, sipXconfig will store the contents in the database and replicate the file as it does in the current code.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14684 ab1d8caa-1f67-47f1-9e81-24633a41865c