XCF-3294 Add ServiceConfigurator and reimplement enforceRole
Service configurator is responsible for properly replicating service,
setting configuration stamp and restarting services.  Previously this
functionality resided in several places which made it difficult to
synchronize it.

Functionality of enforcing the role (starting and stopping services on a
location to make it comply with bunles list) used to reside in
SipxProcessContext. However enforcing role needs to coincide with
generating service configuration and restarting the service - that's why
it is now handled by ServiceConfigurator which can coordinate all the
aspects of starting the service.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14754 ab1d8caa-1f67-47f1-9e81-24633a41865c