XCF-3384: Fixed the problem with locking autocreated conferences.
It ends up that this was a regular expression problem - I was using an
oversimplified expression to extract a conference name - in particular,
I was just using (\w)+. Of course, the word characters don't include other
perfectly valid characters - in this case, the hyphen. Autogenerated user
conferences have a hyphen in their name, e.g. "200-conference". The
hyphen was throwing off the regex and it wasn't recognizing it as an active conference.

To be sure we catch all valid conference names, I'm using SipUriSetting.USER_NAME in
the conference name regular expression.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14778 ab1d8caa-1f67-47f1-9e81-24633a41865c