XX-5764 :  SipXbridge should forward the authentication challenge for challenges in the sipx proxy domain and ITSP originated challenges for "dummy" accounts back to SIPX.
A "dummy ITSP" account is created when forwarding URL calls via
sipxbridge (i.e. using sipxbridge as the SBC) where sipxbridge does
not know the domain of the ITSP apriori. For such a dummy account, when
the outbound call is challenged, the 401/407 should be relayed back to
the call originator.  Similarly when an inbound call is passed through
sipxbridge and sipxbridge sends the call to the sipx proxy server, the
sipx proxy will challenge the call if the From header domain matches
that of sipx. In this case, the 407 from the sipx proxy server needs to
be relayed back to the calling party ( the calling party may have the
credentials to answer the challenge ).

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@15793 ab1d8caa-1f67-47f1-9e81-24633a41865c